requirements.txt
certifi
chardet
click
colorama
dnspython
idna
requests
tld
urllib3
ipwhois